+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.4] - 2026-09-23
+
 ### 🐛 Bug Fixes
 
 - **Codegen: exported types in a generic `<script setup>`** — leading